Fix bug where IDB cursors skipped first element
This was because the `objstore.cursor` function would wait for a callback
to get a handle to the cursor. But that callback would have a value ready,
and we were expecting to call `cursor.continue` each time `cursor.next`
was called! So I made the `objstore.cursor` value get a handle to a
cursor request instead, and the Cursor struct handles calling the
appropriate function to start it off in Cursor.next.